Clean up common methods in iceberg system tables
